pyproject.toml
[tool.poetry]
name = "metasharks"
version = "0.1.0"
description = ""
authors = ["subbotin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
psycopg2-binary = "^2.9.5"
django-multiselectfield = "^0.1.12"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
openpyxl = "^3.0.10"
celery = "^5.2.7"
redis = "^4.3.4"
flower = "^1.2.0"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
